Adjusting to matching field with database

+++ backend/src/main/java/com/swp391_8/schoolhealth/controller/ParentDashboardController.java
@@ -36,7 +36,7 @@ public class ParentDashboardController {
         // If studentId is provided and relevant, the service should handle filtering.
         // For now, assuming notifications are primarily parent-userId based.
         System.out.println("Fetching notifications for parentId: " + parentId + (studentId != null ? ", studentId: " + studentId : ""));
-        List<NotificationDTO> notifications = notificationService.getNotificationsByUserId(parentId);
+        List<NotificationDTO> notifications = notificationService.getNotificationsByUserId(parentId.intValue());
         if (notifications.isEmpty()) {
             return ResponseEntity.noContent().build();
         }
