+++ backend/src/main/java/com/swp391_8/schoolhealth/service/NotificationService.java
@@ -4,6 +4,6 @@ import com.swp391_8.schoolhealth.dto.NotificationDTO;
 import java.util.List;
 
 public interface NotificationService {
-    List<NotificationDTO> getNotificationsByUserId(Long userId);
+    List<NotificationDTO> getNotificationsByUserId(Integer userId);
     // Potentially add methods like markAsRead, createNotification, etc.
 }
